Modify ServeHTTP to handle a third-party having registered a handler.
xhttp.ServeMux is not necessarily the only way to register a handler on its underlying *http.ServeMux.
Therefore, to service a request, it has to take into account the fact that a handler might exist that was not registered through the methods provided by this package.
The introduced change takes this into account.