Reimplement a working version of the gzip Handler.
What was initially uploaded was an ancient version that was not compatible with the changes that have been operated on the xhttp.ServeMux library.
This commit solves this.

The tests have also been updated.
Interesting note: using the Write method of a http.ResponseWriter, content-length and content-type are being sniffed.