Switch execution Context Storer from a sac datastructure to a hmap
Originally, the multiplexer used a small array container object
when defining the execution Context for each request.

Since the hmap package has been created, defining an hybrid between
a regular map[interface{}]interface{} and a sac (unrolled linked list),
it is more sensible to use the latter.

It should be noted that keys are now deep-compared.